+++ src/Bones.php
@@ -241,6 +241,10 @@ class Bones
 
                     $handler->respond($response, $e);
 
+                    if (isset($events)) {
+                        $events->doEvent('bones.end');
+                    }
+
                     return; // Stop iteration
 
                 }
@@ -254,6 +258,10 @@ class Bones
              */
 
             echo '<h1>Error: ' . $e->getMessage() . '</h1>';
+
+            if (isset($events)) {
+                $events->doEvent('bones.end');
+            }
 
         });